+++ operator.ps1
@@ -160,7 +160,7 @@ if ($IsWindows){
 . $pwshSrcDir/Join-While_function.ps1
 . $pwshSrcDir/Join-Until_function.ps1
 . $pwshSrcDir/Join-Step_function.ps1
-. $pwshSrcDir/Join-BlankLine_function.ps1
+. $pwshSrcDir/Join-Blank_function.ps1
 . $pwshSrcDir/kinsoku_function.ps1
 . $pwshSrcDir/filehame_function.ps1
 . $pwshSrcDir/table2md_function.ps1